# docs/swagger.yaml
basePath: /
definitions:
  handlers.AuthRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      whatsapp_number:
        type: string
    type: object
  handlers.AuthResponse:
    properties:
      forge_api_key:
        type: string
      message:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.FileInfo:
    properties:
      name:
        type: string
      size:
        type: integer
      uploaded_at:
        type: string
      url:
        type: string
    type: object
  handlers.ListResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/handlers.FileInfo'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      project:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handlers.ProjectInfo:
    properties:
      file_count:
        type: integer
      name:
        type: string
      total_size:
        type: integer
    type: object
  handlers.ProjectsResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      projects:
        items:
          $ref: '#/definitions/handlers.ProjectInfo'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handlers.UploadResponse:
    properties:
      file:
        type: string
      message:
        type: string
      project:
        type: string
      url:
        type: string
    type: object
  models.File:
    properties:
      id:
        type: string
      mimeType:
        type: string
      name:
        type: string
      path:
        type: string
      projectID:
        type: string
      size:
        type: integer
      uploadedAt:
        type: string
    type: object
  models.Plan:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      storageLimit:
        description: Em bytes
        type: integer
    type: object
  models.Project:
    properties:
      createdAt:
        type: string
      files:
        items:
          $ref: '#/definitions/models.File'
        type: array
      id:
        type: string
      name:
        type: string
      userID:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      forgeAPIKey:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      plan:
        $ref: '#/definitions/models.Plan'
      planID:
        type: string
      projects:
        items:
          $ref: '#/definitions/models.Project'
        type: array
      storageUsage:
        type: integer
      whatsappNumber:
        type: string
    type: object
host: uploader.nativespeak.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a file upload and management API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MidiaForge API
  version: "1.0"
paths:
  /api/delete:
    delete:
      description: Deletes a specific file from a project.
      parameters:
      - description: Project name
        in: query
        name: project
        required: true
        type: string
      - description: File name
        in: query
        name: file
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: File deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: '''project'' and ''file'' parameters are required'
          schema:
            type: string
        "404":
          description: Project not found or File not found
          schema:
            type: string
        "500":
          description: Could not delete file metadata
          schema:
            type: string
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: Delete a file
      tags:
      - api
  /api/list:
    get:
      description: Retrieves a paginated list of files within a specified project
        for the authenticated user.
      parameters:
      - description: Project name
        in: query
        name: project
        required: true
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListResponse'
        "400":
          description: Project name is required
          schema:
            type: string
        "404":
          description: Project not found
          schema:
            type: string
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: List files in a project
      tags:
      - api
  /api/project/delete:
    delete:
      description: Deletes a project that has no files. Projects with files cannot
        be deleted.
      parameters:
      - description: Project name
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Project deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Project name is required or Project has files and cannot be
            deleted
          schema:
            type: string
        "404":
          description: Project not found
          schema:
            type: string
        "500":
          description: Could not delete project
          schema:
            type: string
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: Delete an empty project
      tags:
      - api
  /api/projects:
    get:
      description: Retrieves a paginated list of projects for the authenticated user.
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProjectsResponse'
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: List user's projects
      tags:
      - api
  /api/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a file to a specified project. If the project doesn't exist,
        it will be created.
      parameters:
      - description: Project name
        in: formData
        name: project
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: File uploaded successfully
          schema:
            $ref: '#/definitions/handlers.UploadResponse'
        "400":
          description: 'Bad Request: Error reading file or file is too large'
          schema:
            type: string
        "403":
          description: Storage limit exceeded
          schema:
            type: string
        "413":
          description: File is too large. Max size is 10MB.
          schema:
            type: string
        "415":
          description: 'Invalid file type. Allowed types are: jpeg, png, pdf.'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: Upload a file to a project
      tags:
      - api
  /api/user/rotate-api-key:
    post:
      description: Generates a new API key for the authenticated user, invalidating
        the old one.
      produces:
      - application/json
      responses:
        "200":
          description: 'message: API key rotated successfully, new_api_key: ...'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not rotate API key
          schema:
            type: string
      security:
      - BearerAuth: []
      - APIKeyAuth: []
      summary: Rotate user's API key
      tags:
      - api
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using either email or WhatsApp number and
        password, then returns a JWT token.
      parameters:
      - description: User login credentials (either email or whatsapp_number is required)
        in: body
        name: auth_request
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logged in successfully
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Invalid request body or missing credentials
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
        "500":
          description: Could not generate token
          schema:
            type: string
      summary: Log in a user
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns the user info along with
        an API key.
      parameters:
      - description: User registration details
        in: body
        name: auth_request
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Invalid request body or missing fields
          schema:
            type: string
        "500":
          description: Could not create user or find default plan
          schema:
            type: string
      summary: Register a new user
      tags:
      - auth
schemes:
- https
securityDefinitions:
  APIKeyAuth:
    description: API key for external services.
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"